Refactored flashbar and other core presentation classes #37, #38
