+++ lib/game_board/presentation/core/widgets/game_overview/star_bars_indicator.dart
