+++ lib/game_board/presentation/core/widgets/voice_to_speech_button.dart
@@ -1,4 +1,3 @@
-import 'package:app_settings/app_settings.dart';
 import 'package:avatar_glow/avatar_glow.dart';
 import 'package:flash/flash.dart';
 import 'package:flutter/material.dart';
@@ -6,6 +5,9 @@ import 'package:permission_handler/permission_handler.dart';
 import 'package:speech_to_text/speech_recognition_result.dart';
 import 'package:speech_to_text/speech_to_text.dart' as stt;
 
+import 'flash_widgets.dart';
+
+/// Voice recognitoin widget that only works on Android, and maybe iOS with some configuration changes
 class VoiceToSpeechButton extends StatefulWidget {
   const VoiceToSpeechButton({
     required this.onWordHeard,
@@ -47,40 +49,10 @@ class _VoiceToSpeechButtonState extends State<VoiceToSpeechButton> {
                 showFlash(
                     context: context,
                     builder: (context, controller) {
-                      return Flash(
-                          controller: controller,
-                          behavior: FlashBehavior.floating,
-                          position: FlashPosition.bottom,
-                          barrierBlur: 2,
-                          boxShadows: const [
-                            BoxShadow(
-                                color: Colors.black12,
-                                offset: Offset(0, 1),
-                                blurRadius: 10,
-                                spreadRadius: 5)
-                          ],
-                          child: FlashBar(
-                            title: Text(
-                              "No hay permisos suficientes",
-                              style: Theme.of(context).textTheme.headline6,
-                            ),
-                            content: const Text(
-                                "Se necesita permisos de microfono para poder utilizar el reconocimiento por voz. Pulse el icono del engranaje para ir a los ajustes de la aplicación y cambie los permisos"),
-                            primaryAction: IconButton(
-                              onPressed: () async {
-                                await AppSettings.openAppSettings();
-                              },
-                              icon: const Icon(Icons.settings_applications),
-                              iconSize: 36,
-                              color: Colors.black,
-                            ),
-                            icon: Icon(
-                              Icons.warning,
-                              color: Colors.amberAccent.shade700,
-                            ),
-                          ));
+                      return MicrophonePermissionFlash(
+                        controller: controller,
+                      );
                     });
-                // _showB;
                 return;
               }
               await speech.initialize();
