+++ lib/game_board/presentation/games_browse/widgets/body/games_browse_sliver_list.dart
@@ -9,9 +9,9 @@ import '../../../../../constants/app_margins_and_sizes.dart';
 import '../../../../../core/optional.dart';
 import '../../../../zacatrus/domain/browse_page/game_overview.dart';
 import '../../../core/routing/games_router_delegate.dart';
-import '../../../core/widgets/price.dart';
-import '../../../core/widgets/reviews_number.dart';
-import '../../../core/widgets/star_bars_indicator.dart';
+import '../../../core/widgets/game_overview/price.dart';
+import '../../../core/widgets/game_overview/reviews_number.dart';
+import '../../../core/widgets/game_overview/star_bars_indicator.dart';
 
 class GamesBrowseSliverList extends StatelessWidget {
   const GamesBrowseSliverList({
